Reshaped the code after review #6504
